desktop_app/requirements.txt
altgraph==0.17.4
blinker==1.9.0
bottle==0.13.4
cffi==1.17.1
click==8.2.1
clr_loader==0.2.7.post0
colorama==0.4.6
cryptography==45.0.7
Flask==3.1.2
flask-cors==6.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==25.0
pefile==2023.2.7
pillow==11.3.0
proxy_tools==0.1.0
psutil==6.1.1
pycparser==2.22
pyinstaller==6.15.0
pyinstaller-hooks-contrib==2025.8
pystray==0.19.5
pythonnet==3.0.5
pywebview==6.0
pywin32-ctypes==0.2.3
six==1.17.0
typing_extensions==4.15.0
waitress==3.0.2
Werkzeug==3.1.3
